Nabil's Assesment
Retrieve the data from the mock API. Done
Output the data in a list, including properties from the data that are appropriate for a list view. Done
Implement a category filter - this can be single or multi-select. Done
Implement pagination - this can be traditional numbered pages or "load more". Done
Use semantic markup where possible. Done
Create a responsive layout with HTML and CSS. Done
Convert the application to use TypeScript instead of JavaScript. Done
Use a CSS preprocessor or CSS-in-JS rather than plain CSS. Done